update long / short exposure values and charts
